requirement.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blessed==1.19.1
brotlipy==0.7.0
cachetools==5.2.0
captum==0.5.0
certifi @ file:///opt/conda/conda-bld/certifi_1663615672595/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.2.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cycler==0.11.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
detectron2==0.1.3+cu101
docker-pycreds==0.4.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
filelock==3.7.1
fonttools==4.34.4
frozenlist==1.3.1
fsspec==2022.7.1
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20220512
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.10.0
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.47.0
huggingface-hub==0.8.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.21.1
importlib-metadata==4.12.0
install==1.3.5
iopath==0.1.10
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646983213839/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///tmp/abs_94gruux8u8/croots/recipe/ipython_1659529858706/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
joblib==1.1.0
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mmcv-full==1.6.1
mock==4.0.3
multidict==6.0.2
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847911453/work
networkx==2.6.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nvidia-ml-py==11.495.46
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
portalocker==2.5.1
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
psutil==5.9.2
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.1
pydot==1.4.2
pyemd==0.5.1
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==1.5.8
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
regex==2022.7.25
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.0
sentry-sdk==1.9.2
setproctitle==1.3.1
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.0.0
smmap==5.0.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.12.1
torch==1.2.0
torchaudio==0.8.0
torchmetrics==0.9.3
torchvision==0.4.0
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
transformers @ file:///home/pr03/jh_pjh/xAI/ICASSP2023/OFA/transformers
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
wandb==0.13.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1

